Remove autofix workflow due to GitHub limitations
In order to have checks run on commits created by CI/CD, you must have a
personal access token. Well, you have to store that token as a secret.
When branches that are not part of the repo (forks) come in, they run
without secrets, so you can't get the token.